requirements.txt
Pillow==9.0.1
streamlit==1.7.0
yfinance==0.1.70
